add reject buttons
