+++ client/containers/Pub/PubHeader/DownloadButton.tsx
@@ -1,22 +1,19 @@
 import React from 'react';
-import { Callout } from '@blueprintjs/core';
+import { Button, Callout, Classes } from '@blueprintjs/core';
+import { EditorState, Transaction } from 'prosemirror-state';
 
 import { PopoverButton } from 'components';
 import { hasSuggestions } from 'client/utils/suggestedEdits';
 import { pubData } from 'utils/storybook/data';
 import { usePageContext } from 'utils/hooks';
+import {
+	acceptSuggestions,
+	rejectSuggestions,
+} from 'components/Editor/plugins/suggestedEdits/resolve';
 
 import Download from './Download';
 import { usePubContext } from '../pubHooks';
 import SmallHeaderButton from './SmallHeaderButton';
-
-const PopoverContent = () => {
-	return (
-		<Callout className="text-info" intent="warning">
-			You still have pending edits. Resolve your edits before downloading this Pub.
-		</Callout>
-	);
-};
 
 const ConditionalWrapper = ({ condition, wrapper, children }) =>
 	condition ? wrapper(children) : children;
@@ -25,6 +22,50 @@ const DowloadButton = () => {
 	const { featureFlags } = usePageContext();
 	const { collabData } = usePubContext();
 	const { editorChangeObject } = collabData;
+
+	const suggestedEditsAction = (action: {
+		(state: EditorState, from: number, to: number): Transaction;
+	}) => {
+		if (!hasSuggestions(editorChangeObject)) return;
+		if (!editorChangeObject) return;
+		const editorState = editorChangeObject.view.state;
+		const size = editorChangeObject.view.state.doc.nodeSize;
+		const tr = action(editorState, 0, size - 2);
+		editorChangeObject.view.dispatch(tr);
+	};
+
+	const handleSuggestedEditsAccept = () => {
+		suggestedEditsAction(acceptSuggestions);
+	};
+
+	const handleSuggestedEditsReject = () => {
+		suggestedEditsAction(rejectSuggestions);
+	};
+
+	const PopoverContent = () => {
+		return (
+			<React.Fragment>
+				<Callout className="text-info" intent="warning">
+					<p>
+						You still have pending edits. Resolve your edits before downloading this
+						Pub.
+					</p>
+					<div className={Classes.DIALOG_FOOTER_ACTIONS}>
+						<Button
+							text="Reject All"
+							intent="danger"
+							onClick={handleSuggestedEditsReject}
+						/>
+						<Button
+							text="Accept All"
+							intent="success"
+							onClick={handleSuggestedEditsAccept}
+						/>
+					</div>
+				</Callout>
+			</React.Fragment>
+		);
+	};
 	return (
 		<ConditionalWrapper
 			condition={hasSuggestions(editorChangeObject) && featureFlags.suggestedEdits}
